+++ resources/js/Components/Global/Ott/Elements/Playlist.vue
@@ -89,9 +89,10 @@ const shouldDisplayOtt = computed(() => {
 })
 
 const showUpgrade = computed(() => {
-  return !userStore.isSubscriber &&
-      !userStore.isVip &&
-      !userStore.isAdmin;
+  return false
+  // return !userStore.isSubscriber &&
+  //     !userStore.isVip &&
+  //     !userStore.isAdmin;
 });
 
 </script>